feat(proxy/backups): move fetchPartitionFiles implementation to RemoteAdapter
